+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.15.1] - (08.06.2024)
+
+## Changes
+
+- added compare mode to the encharge backup level sensors
+- config schema updated
+
 ## [7.15.0] - (08.06.2024)
 
 ## Changes